refactor(query-builder): extract database types into separate module for improved modularity
- Create new `database.types.ts` file to separate type definitions
- Move complex type interfaces and type utility functions from `database.ts`
- Update import statements across project to use new type module
- Improve code organization and type management
- Enhance type reusability and maintainability
