--- packages/query-builder/src/__tests__/fixtures/models/base-schema.ts
@@ -1,14 +1,0 @@
-export interface TestDatabase {
-	users: {
-		id: number;
-		name: string;
-		email: string;
-		slug: string;
-		status: string;
-		created_at: Date;
-	};
-	user_favorites: {
-		user_id: number;
-		article_id: string;
-	};
-}
