+++ packages/query-builder/src/__tests__/integrations/crud.spec.ts
@@ -1,11 +1,6 @@
 import { afterEach, beforeEach, describe, expect, it } from 'vitest';
 import { createModel } from '../../model';
-import {
-	DB,
-	cleanupDatabase,
-	db,
-	initializeDatabase,
-} from '../fixtures/migration';
+import { cleanupDatabase, db, initializeDatabase } from '../fixtures/migration';
 
 // Helper function to generate IDs
 const generateId = (): string => {
