+++ packages/query-builder/src/__tests__/integrations/error-handling.spec.ts
@@ -1,7 +1,6 @@
-import { sql } from 'kysely';
 import { afterEach, beforeEach, describe, expect, it } from 'vitest';
 import { createModel } from '../../model';
-import type { DB, Timestamp, Users } from '../fixtures/migration';
+import type { Timestamp, Users } from '../fixtures/migration';
 import {
 	cleanupDatabase,
 	db,
