+++ packages/query-builder/src/__tests__/unit/advanced-operations.spec.ts
@@ -1,14 +1,9 @@
 import { sql } from 'kysely';
-import type { UpdateResult } from 'kysely';
 import { describe, expect, it, vi } from 'vitest';
-import type { Database } from '../../database';
-import { createModel } from '../../model';
 import {
 	type MockExpressionBuilder,
-	type TestMockDatabase,
 	createMockDatabase,
 	createMockExpressionBuilder,
-	createMockReturnThis,
 } from '../fixtures/mock-db';
 
 // Define test database types
