+++ packages/query-builder/src/__tests__/unit/mixins/assign.spec.ts
@@ -1,6 +1,6 @@
 import type { Selectable } from 'kysely';
-import { beforeEach, describe, expect, it, vi } from 'vitest';
-import withAssign, { ModelWithAssign } from '../../../mixins/assign';
+import { beforeEach, describe, expect, it } from 'vitest';
+import withAssign from '../../../mixins/assign';
 import type { ModelFunctions } from '../../../model';
 import { DB } from '../../fixtures/migration';
